+++ CHANGELOG.md
@@ -10,9 +10,10 @@ All notable changes to the package will be documented here.
 ## [4.27.xx] - 2025-12-xx [Future]
 
 - **Added** *new* utility `parseMs` to convert any time value to *milliseconds* with new type guard `isTimeWithUnit`.
-- **Added** `new` class `Cipher` to *encrypt/decrypt* string with *secret*.
-- **Added** `new` class `SimpleToken` to *generate, decode* and *verify* token like `JWT`.
+- **Added** *new* class `Cipher` to *encrypt/decrypt* string with *secret*.
+- **Added** *new* class `SimpleToken` to *generate, decode* and *verify* token like `JWT`.
 - **Updated** `sha1` *encoding algorithm*. Now it avoids depending on `TextEncoder`.
+- **Added** *new* `JSON` utilities `stableStringify` for stable, deterministic stringifying and `stripJsonEdgeGarbage` to strip `JSON` string.
 -->
 
 ## [4.27.11] - 2025-11-29